cosolidate code
